+++ deepinv/models/equivariant.py
@@ -1,6 +1,7 @@
+from typing import Optional
 import numpy as np
-
 import torch
+from deepinv.transform import Transform, Rotate, Reflect
 
 
 class EquivariantDenoiser(torch.nn.Module):
@@ -23,113 +24,52 @@ class EquivariantDenoiser(torch.nn.Module):
     .. math::
         \operatorname{D}^{\text{MC}}_{\sigma}(x) = T_g^{-1}(\operatorname{D}_{\sigma}(T_g(x))).
 
+    .. note::
+
+        We have implemented many popular geometric transforms, see :ref:`docs <transform>`. You can set the number of Monte Carlo samples by passing ``n_trans``
+        into the transforms, for example ``Rotate(n_trans=2)`` will average over 2 samples per call. For rotate and reflect, by setting ``n_trans``
+        to the maximum (e.g. 4 for 90 degree rotations, 2 for 1D reflections), it will average over the whole group, for example:
+
+        ``Rotate(n_trans=4, multiples=90, positive=True) * Reflect(n_trans=2, dims=[-1])``
+
+    See :ref:`sphx_glr_auto_examples_basics_demo_transforms.py` for an example.
 
     :param callable denoiser: Denoiser :math:`\operatorname{D}_{\sigma}`.
-    :param str transform: type of geometric transformation. Can be either 'rotations', 'flips' or 'rotoflips'.
-        If 'rotations', the group of transformations contains the 4 rotations by multiples of 90 degrees; if 'flips',
-        the group of transformations contains the 2 horizontal and vertical flips; if 'rotoflips', the group of
-        transformations contains the 8 rotations and flips.
-    :param bool random: if True, the denoiser is applied to a randomly transformed version of the input image.
+    :param Transform transform: geometric transformation. If None, defaults to rotations of multiples of 90 with horizontal flips (see note above).
+        See :ref:`docs <transform>` for list of available transforms.
+    :param bool random: if True, the denoiser is applied to a randomly transformed version of the input image
+        each time i.e. a Monte-Carlo approximation of an equivariant denoiser.
         If False, the denoiser is applied to the average of all the transformed images, turning the denoiser into an
-        equivariant denoiser with respect to the chosen group of transformations. Otherwise, it is a Monte-Carlo
-        approximation of an equivariant denoiser.
+        equivariant denoiser with respect to the chosen group of transformations. Ignored if ``transform`` is provided.
     """
 
-    def __init__(self, denoiser, transform="rotations", random=True):
+    def __init__(self, denoiser, transform: Optional[Transform] = None, random=True):
         super().__init__()
         self.denoiser = denoiser
-        self.rotations = True if "rot" in transform else False
-        self.flips = True if "flip" in transform else False
-        self.random = random
 
-    def forward(self, x, sigma):
+        if transform is not None:
+            self.transform = transform
+        else:
+            if random:
+                self.transform = Rotate(
+                    n_trans=1, multiples=90, positive=True
+                ) * Reflect(n_trans=1, dims=[-1])
+            else:
+                self.transform = Rotate(
+                    n_trans=4, multiples=90, positive=True
+                ) * Reflect(n_trans=2, dims=[-1])
+
+    def forward(self, x, *denoiser_args, **denoiser_kwargs):
         r"""
-        Applies the denoiser to the input image with the appropriate transformation.
+        Symmetrize the denoiser by the transformation to create an equivariant denoiser and apply to input.
+
+        The symmetrization collects the average if multiple samples are used (controlled with ``n_trans`` in the transform).
 
         :param torch.Tensor x: input image.
-        :param float sigma: noise level.
+        :param \*denoiser_args: args for denoiser function e.g. sigma noise level.
+        :param \**denoiser_kwargs: kwargs for denoiser function e.g. sigma noise level.
         :return: denoised image.
         """
-        return denoise_rotate(
-            self.denoiser,
-            x,
-            sigma,
-            rotations=self.rotations,
-            flips=self.flips,
-            random=self.random,
+        return self.transform.symmetrize(self.denoiser, average=True)(
+            x, *denoiser_args, **denoiser_kwargs
         )
-
-
-def denoise_rotate(
-    denoiser,
-    image,
-    sigma,
-    rotations=True,
-    flips=False,
-    random=True,
-):
-    r"""
-    Applies a geometric transform (rotations and/or flips) to the input image, denoises it with the denoiser and
-    transform back the result. The output is either the average of all the transformed images (if random=False) or a
-    randomly transformed version of the denoised image (if random=True).
-
-    :param callable denoiser: Denoiser :math:`\operatorname{D}_{\sigma}`.
-    :param torch.Tensor image: input image.
-    :param float sigma: noise level.
-    :param bool rotations: if True, rotations are applied to the input image.
-    :param bool flips: if True, flips are applied to the input image.
-    :param bool random: if True, the denoiser is applied to a randomly transformed version of the input image.
-    :return: denoised image.
-    """
-    if random:
-        if rotations:
-            idx = np.random.randint(8) if flips else np.random.randint(4)
-        elif flips:
-            idx = np.random.choice([4, 6])
-        denoised = denoise_rotate_flip_fn(denoiser, image, sigma, idx)
-    else:
-        if rotations:
-            list_idx = list(range(8)) if flips else list(range(4))
-        elif flips:
-            list_idx = [4, 6]
-        denoised = torch.zeros_like(image)
-        for idx in list_idx:
-            denoised = denoised + denoise_rotate_flip_fn(denoiser, image, sigma, idx)
-        denoised = denoised / len(list_idx)
-    return denoised
-
-
-def denoise_rotate_flip_fn(denoiser, x, sigma_den, idx):
-    if idx == 0:
-        out = denoiser(x, sigma_den)
-    elif idx == 1:
-        out = rot3(denoiser(rot1(x), sigma_den))
-    elif idx == 2:
-        out = rot2(denoiser(rot2(x), sigma_den))
-    elif idx == 3:
-        out = rot1(denoiser(rot3(x), sigma_den))
-    elif idx == 4:
-        out = hflip(denoiser(hflip(x), sigma_den))
-    elif idx == 5:
-        out = hflip(rot3(denoiser(rot1(hflip(x)), sigma_den)))
-    elif idx == 6:
-        out = hflip(rot2(denoiser(rot2(hflip(x)), sigma_den)))
-    elif idx == 7:
-        out = hflip(rot1(denoiser(rot3(hflip(x)), sigma_den)))
-    return out
-
-
-def hflip(x):
-    return torch.flip(x, dims=[-1])
-
-
-def rot1(x):
-    return torch.rot90(x, k=1, dims=[-2, -1])
-
-
-def rot2(x):
-    return torch.rot90(x, k=2, dims=[-2, -1])
-
-
-def rot3(x):
-    return torch.rot90(x, k=3, dims=[-2, -1])