+++ deepinv/optim/prior.py
@@ -413,7 +413,8 @@ class TVPrior(Prior):
         :param torch.Tensor x: Variable :math:`x` at which the prior is computed.
         :return: (torch.Tensor) prior :math:`g(x)`.
         """
-        return torch.sum(torch.sqrt(torch.sum(self.nabla(x) ** 2, axis=-1)))
+        y = torch.sqrt(torch.sum(self.nabla(x) ** 2, dim=-1))
+        return torch.sum(y.reshape(x.shape[0], -1), dim=-1)
 
     def prox(self, x, *args, gamma=1.0, **kwargs):
         r"""Compute the proximity operator of TV with the denoiser :class:`~deepinv.models.TVDenoiser`.
@@ -583,3 +584,74 @@ class PatchNR(Prior):
         latent_x, logdet = self.normalizing_flow(x.view(B * n_patches, -1))
         logpz = 0.5 * torch.sum(latent_x.view(B, n_patches, -1) ** 2, -1)
         return logpz - logdet.view(B, n_patches)
+
+
+class L12Prior(Prior):
+    r"""
+    :math:`\ell_{1,2}` prior :math:`\reg{x} = \sum_i\| x_i \|_2`.
+    The :math:`\ell_2` norm is computed over a tensor axis that can be defined by the user. By default, ``l2_axis=-1``.
+    |sep|
+
+    :Examples:
+    >>> import torch
+    >>> from deepinv.optim import L12Prior
+    >>> seed = torch.manual_seed(0) # Random seed for reproducibility
+    >>> x = torch.randn(2, 1, 3, 3) # Define random 3x3 image
+    >>> prior = L12Prior()
+    >>> prior.g(x)
+    tensor([5.4949, 4.3881])
+    >>> prior.prox(x)
+    tensor([[[[-0.4666, -0.4776,  0.2348],
+              [ 0.3636,  0.2744, -0.7125],
+              [-0.1655,  0.8986,  0.2270]]],
+    <BLANKLINE>
+    <BLANKLINE>
+            [[[-0.0000, -0.0000,  0.0000],
+              [ 0.7883,  0.9000,  0.5369],
+              [-0.3695,  0.4081,  0.5513]]]])
+
+    """
+
+    def __init__(self, *args, l2_axis=-1, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.explicit_prior = True
+        self.l2_axis = l2_axis
+
+    def g(self, x, *args, **kwargs):
+        r"""
+        Computes the regularizer :math:`\reg{x} = \sum_i\| x_i \|_2`.
+
+        :param torch.Tensor x: Variable :math:`x` at which the prior is computed.
+        :return: (torch.Tensor) prior :math:`\reg{x}`.
+        """
+        x_l2 = torch.norm(x, p=2, dim=self.l2_axis)
+        return torch.norm(x_l2.reshape(x.shape[0], -1), p=1, dim=-1)
+
+    def prox(self, x, *args, gamma=1.0, **kwargs):
+        r"""
+        Calculates the proximity operator of the :math:`\ell_{1,2}` function at :math:`x`.
+
+        More precisely, it computes
+
+        .. math::
+            \operatorname{prox}_{\gamma g}(x) = (1 - \frac{\gamma}{max{\Vert x \Vert_2,\gamma}}) x
+
+
+        where :math:`\gamma` is a stepsize.
+
+        :param torch.Tensor x: Variable :math:`x` at which the proximity operator is computed.
+        :param float gamma: stepsize of the proximity operator.
+        :param int l2_axis: axis in which the l2 norm is computed.
+        :return torch.Tensor: proximity operator at :math:`x`.
+        """
+
+        tau_gamma = torch.tensor(gamma)
+
+        z = torch.norm(x, p=2, dim=self.l2_axis, keepdim=True)
+        # Creating a mask to avoid diving by zero
+        # if an element of z is zero, then it is zero in x, therefore torch.multiply(z, x) is zero as well
+        mask_z = z > 0
+        z[mask_z] = torch.max(z[mask_z], tau_gamma)
+        z[mask_z] = torch.tensor(1.0) - tau_gamma / z[mask_z]
+
+        return torch.multiply(z, x)