+++ deepinv/transform/__init__.py
@@ -3,3 +3,6 @@ from .shift import Shift
 from .scale import Scale
 from .base import Transform
 from .projective import Homography
+from .reflect import Reflect
+from .temporal import ShiftTime
+from .diffeomorphism import CPADiffeomorphism